+++ pkgs/stack_trace/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.12.2-dev
+
+* Relax parsing of V8 traces to allow a missing initial description.
+
 ## 1.12.1
 
 * Move to `dart-lang/tools` monorepo.